Create a file "backend" that contains all non-UI code.
